requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file base.txt base.in
#
certifi==2016.8.2         # via python-telegram-bot
future==0.15.2            # via python-telegram-bot
python-telegram-bot==5.0.0
pytz==2016.6.1
requests==2.10.0          # via vk
SQLAlchemy==1.0.14
urllib3==1.16             # via python-telegram-bot
vk==2.0.2
